Chore/update project references and bump package versions
