--- AlexaVoxCraft.TestKit/AlexaVoxCraft.TestKit.csproj
@@ -1,33 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-    <PropertyGroup>
-        <ImplicitUsings>enable</ImplicitUsings>
-        <Nullable>enable</Nullable>
-        <TargetFrameworks>net8.0;net9.0;net10.0</TargetFrameworks>
-        <LangVersion>default</LangVersion>
-        <IsPackable>false</IsPackable>
-        <TestingPlatformDotnetTestSupport>true</TestingPlatformDotnetTestSupport>
-        <UseMicrosoftTestingPlatformRunner>true</UseMicrosoftTestingPlatformRunner>
-    </PropertyGroup>
-
-    <ItemGroup>
-      <PackageReference Include="AutoFixture" Version="4.18.1" />
-      <PackageReference Include="AutoFixture.AutoNSubstitute" Version="4.18.1" />
-      <PackageReference Include="AutoFixture.Xunit3" Version="4.19.0" />
-      <PackageReference Include="AwesomeAssertions" Version="9.3.0" />
-      <PackageReference Include="NSubstitute" Version="5.3.0" />
-        <PackageReference Include="xunit.v3" Version="3.2.1" />
-        <PackageReference Include="xunit.runner.visualstudio" Version="3.1.5">
-            <PrivateAssets>all</PrivateAssets>
-            <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
-        </PackageReference>
-        <PackageReference Include="Microsoft.NET.Test.Sdk" Version="18.0.1" />
-    </ItemGroup>
-
-    <ItemGroup>
-      <ProjectReference Include="..\src\AlexaVoxCraft.MediatR.Lambda\AlexaVoxCraft.MediatR.Lambda.csproj" />
-      <ProjectReference Include="..\src\AlexaVoxCraft.MediatR\AlexaVoxCraft.MediatR.csproj" />
-      <ProjectReference Include="..\src\AlexaVoxCraft.Model\AlexaVoxCraft.Model.csproj" />
-    </ItemGroup>
-
-</Project>
