+++ samples/Sample.Host.Function/Sample.Host.Function.csproj
@@ -28,7 +28,7 @@
         <CompilerVisibleProperty Include="EnableMediatRGeneratorInterceptor"/>
     </ItemGroup>
     <ItemGroup>
-        <PackageReference Include="LayeredCraft.Logging.CompactJsonFormatter" Version="1.0.1.3"/>
+        <PackageReference Include="LayeredCraft.Logging.CompactJsonFormatter" Version="1.0.2.4" />
         <PackageReference Include="Serilog.AspNetCore" Version="10.0.0"/>
         <PackageReference Include="Serilog.Sinks.Console" Version="6.1.1"/>
         <RuntimeHostConfigurationOption Include="System.Globalization.AppLocalIcu" Value="72.1.0.3"/>
