+++ src/AlexaVoxCraft.MediatR.Lambda/AlexaVoxCraft.MediatR.Lambda.csproj
@@ -15,7 +15,7 @@
     <ItemGroup>
         <PackageReference Include="Amazon.Lambda.RuntimeSupport" Version="1.14.1" />
         <PackageReference Include="Amazon.Lambda.Serialization.SystemTextJson" Version="2.4.4"/>
-        <PackageReference Include="LayeredCraft.Logging.CompactJsonFormatter" Version="1.0.1.3" />
+        <PackageReference Include="LayeredCraft.Logging.CompactJsonFormatter" Version="1.0.2.4" />
         <PackageReference Include="Serilog.AspNetCore" Version="9.0.0"/>
         <PackageReference Include="Serilog.Enrichers.Activity" Version="1.0.0.3" />
     </ItemGroup>
