+++ src/AlexaVoxCraft.MinimalLambda/AlexaVoxCraft.MinimalLambda.csproj
@@ -18,7 +18,7 @@
     </ItemGroup>
 
     <ItemGroup>
-      <PackageReference Include="MinimalLambda" Version="2.0.0" />
+      <PackageReference Include="MinimalLambda" Version="2.1.1" />
     </ItemGroup>
 
 </Project>
