+++ test/AlexaVoxCraft.TestKit/Attributes/MediatRLambdaAutoDataAttribute.cs
