+++ test/AlexaVoxCraft.TestKit/RequestSpecifications/BaseTypedSpecification.cs
