+++ test/AlexaVoxCraft.TestKit/RequestSpecifications/LogEventSpecification.cs
