+++ test/AlexaVoxCraft.TestKit/RequestSpecifications/SkillRequestFactorySpecification.cs
@@ -1,0 +1,7 @@
+using AlexaVoxCraft.MediatR;
+
+namespace AlexaVoxCraft.TestKit.RequestSpecifications;
+
+public class SkillRequestFactorySpecification : BaseTypedSpecification<SkillRequestFactory>
+{
+}
