+++ test/AlexaVoxCraft.TestKit/RequestSpecifications/SkillResponsetSpecification.cs
@@ -1,0 +1,7 @@
+using AlexaVoxCraft.Model.Response;
+
+namespace AlexaVoxCraft.TestKit.RequestSpecifications;
+
+public class SkillResponseSpecification : BaseTypedSpecification<SkillResponse>
+{
+}
