+++ test/AlexaVoxCraft.TestKit/SpecimenBuilders/LogEventSpecimenBuilder.cs
